move stuff to room handler
